Allow pinning of the co-located database (#306)
This PR removes the `limit_app_cpus` functionality from the colocated
database. On deeper inspection, this was not functioning as expected
on multiple workfoad manager. The database was being pinned to the
first N logical CPUs on the machine (without the user being able to
change the pinning) and when `limit_app_cpus` was True, only the
launch command and not the launched applications was being pinned.
Additionally, any pinning that the user was trying to do for their Slurm-
launched application, was being silently removed. Both these factors
led to significant performance degradation especially on Cray EX-like
platforms.

To solve this problem, users can now set a `custom_pinning` argument
for the co-located database. By default on Linux, the database will still
be assigned to the first N CPUs, however for machines like Frontier
which reserve some CPUs for system processes, or for advanced users
who might want to pin the DB CPUs to a NUMA node and/or by proximity
to GPUs can now pass in a list of CPU ids for the database to be
assigned.

[ committed by @ashao ]
[ reviewed by @MattToast and @mellis13 ]