Simplify the logic in QsubBatchSettings
Much of the need to check the type and value of the the nodes
property in QsubBatchSettings is because there are two technically
valid, but not quite equivalent ways of setting the number of
nodes. Now, we check at various points that the both 'select' and
'nodes' is not set. Additionally, both routes can be used to set
the internal _nodes property if it needs to be accessed within
Python