Improve testing robustness (#381)
A number of defects were found in the testing that primarily affected
the tests run on HPC platforms. Many of these were uncovered during
the recent introductions of major features that touched on various
aspects of the testing suite. Most fixes are focused primarily on changes
to the actual tests as opposed to fundamental changes in the underlying
code base. Some of the major changes are:
- Most tests are run in their own experiment directory avoiding the
  overwrite of directories between tests
- When attempting to run multi-gpu tests, a bug (presumably in RedisAI)
   was found that prevents the setting of multiple GPUs when using the
   Tensorflow backend. These tests now only use a single GPU regardless
   of the value of `SMARTSIM_TEST_NUM_GPUS`
- Ensures that tests that spin up an `Orchestrator` always stop it before
   exiting, either due to success or failure of a different component of the
   test

Lastly, changes were also made to `QsubBatchSettings` to add support
for PBS-like platforms that use the `resources` tag to define additional
resources and/or otherwise customize PBS batch jobs

[ committed by @al-rigazzi and @ashao ]
[ reviewed by: @MattToast ]

Co-authored-by: Alessandro Rigazzi <al.rigazzi@hpe.com>
Co-authored-by: Andrew Shao <andrew.shao@hpe.com>