Update documentation (#411)
This PR adds a section named ML Features to the documentation. The section contains multiple examples of how ML models can be uploaded to and executed on the database. The PR also adds a section to the Online Analysis tutorial. The section uses TorchScript functions to post-process the simulation data.

[ committed by @al-rigazzi ]
[ reviewed by @amandarichardsonn @mellis13 ]