SmartSim Experiment docs (#422)
This PR merges the Experiment.rst file into a branch that will
encapsulate the entirety of the SmartSim documentation refactor. The
Experiment.rst file offers documentation for an Experiment, a Launcher,
the Experiment entities and an example demonstrating initializing,
starting and stopping from within an experiment. The documentation is
aimed to be accessible.

[reviewed by @mellis13 @ashao @billschereriii @juliaputko ]
[committed by @amandarichardsonn ]