Add support for Mac OSX on Apple Silicon (#465)
With the new ml_lib_builder repository we can now ship a version of
libtorch that is compiled for Mac OSX on Apple Silicon (arm64). Here
the RedisAIBuilder method is reworked to detect whether this
platform and retrieve the appropriate version of libtorch. Some
additional refactoring was done to improve the internals of this class.

[ committed by @ashao ]
[ reviewed by @MattToast ]