Avoid setting `task_id` on a managed step in telemetry monitor (#557)
Ensures that a managed step-mapping doesn't include a `task_id`.

The telemetry monitor exhibits a defect where it logs errors from a task
manager, even with only managed tasks being monitored for updates:

![image](https://github.com/CrayLabs/SmartSim/assets/3595025/84921e5b-144b-4fcd-8289-48d2504deaac)

This fix modifies the telemetry monitor to not set a `task_id` when
adding items to the `step_mapping` collection. This avoids triggering
lookups for unmanaged processes.

![image](https://github.com/CrayLabs/SmartSim/assets/3595025/dfc7cfad-a875-45b3-91d2-fa19407c9d0c)

[ committed by @ankona ]
[ approved by @MattToast ]