Dragon launcher (#580)
This PR adds a Dragon-based launcher to SmartSim. The new launcher can be selected by specifying `launcher="dragon"`, e.g. in the `Experiment` constructor. The Dragon launcher can be used on systems where the PBS or Slurm schedulers are available (MPI-based applications currently require Cray PMI or Cray PALS). 
A new `--dragon` option was added to `smart build` to install the correct Dragon package in the Python environment. More information about the Dragon launcher is available in the dedicated documentation section. Please note that the Dragon launcher is at an early development stage, and we would love to hear feedback about it.

[ committed by @al-rigazzi @ankona @MattToast @amandarichardsonn ]
[ reviewed by @mellis13 @ankona @MattToast @amandarichardsonn ]

---------

Co-authored-by: Matt Drozt <matthew.drozt@gmail.com>
Co-authored-by: Chris McBride <3595025+ankona@users.noreply.github.com>