Adapt tests to reuse Orchestrator (#567)
Tests which needed to launch an Orchestrator were spinning up and
shutting down their own instances. This led to a number of cases where a
single test failing would cascade into failures of other tests.
Additionally, this also meant that a significant amount of time in the
tests was spent waiting for Orchestrators to launch.

This PR adds a session-scoped fixture that returns an Orchestrator. Most
tests which use an Orchestrator have been updated to use this fixture;
the remaining for various reasons still need to spin up their own (for
example the multiple database tests need to have a named Orchestrator).

[ committed by @ashao ]
[ reviewed by @ankona @AlyssaCote ]

Co-authored-by: Matt Drozt <drozt@hpe.com>