Building SmartSim without ML backends (#601)
Fix an error that would prevent ``smart build`` from moving a successfully
compiled RedisAI shared object to the install location expected by SmartSim
if no ML backend installations were found.

The reason is that after RedisAI is built using the `smart` tool, the
resulting library is only installed to the `lib` folder if and only if
the folder `backends/` exists, which it does not if no backends are
installed. Since after this step the original build folder is deleted
and with it the compiled library.

This problem does not occur if any of the backends (TF, PT, ONNX) is
installed. However, since they are not needed for many applications it
would additional complications and effort to compile them if not
necessary. Also compiling RedisAI by itself on pointing `RAI_PATH` to
the installation also works, but poses additional effort.

To circumvent this problem this change will install the
RedisAI library by itself if it was built.

[ committed by @m-kurz ]
[ reviewed by @MattToast ]