Implement support for SGE (#610)
SGE shares some similarities to PBS/Torque-like launchers, but the
differences are significant enough to warrant their own separate
implementations. Notably, SGE has a qacct utility (similar to SLURM's
sacct) to query for the historical record of a job. Additionally, unique
amongst the launchers, SGE does not allow a way for a user to specify
the number of nodes needed and requires the user to select a
admin-configured parallel environment profile (e.g. mpi or smp).

The changes add a new SGE launcher and SGEQsubBatchSettings to 
enable support for SmartSim applications on SGE machines.

[ committed by @ashao ]
[ reviewed by @al-rigazzi ]