More easily discoverable dependencies (#635)
setup.py used to define dependencies in a way that was not amenable to
code scanning tools. Direct dependencies now appear directly in the
setup call and the definition of the SmartRedis version has been
removed. Additionally, the code scanning tool was failing to detect some
of the dependencies due to the existence of the requirements-doc.txt
file. These requirements are now listed in the `docs` extra.

[ committed by @ashao ]
[ reviewed by @ankona ]