Dispatch Jobs to Launchers based on Settings (#620)
Create a `Dispatcher` class and a global `DEFAULT_DISPATCHER` instance.
This can be used by the experiment class to get which type of launcher
are to be used for a specific `LaunchSettings` by maintaining an
internal mapping of arguments type -> launcher type. This mapping can be
built in a type safe way by calling `Dispatcher.disptach`. Experiments
now use the `DEFAULT_DISPATCHER` to start jobs through an 
`Experiment.start` method

[ committed by @MattToast ]
[ reviewed by @amandarichardsonn @mellis13 ]