Enable specification of target hostname for a dragon task (#660)
## Description

This PR adds two features:

1. Ability to specify hostnames that tasks should run on
2. Enable tasks colocation

### Specifying Hostnames

The existing `DragonRunRequest` supported the ability to specify a
hostname when creating a policy used to run a task. However, the
hostnames were not exposed to clients.

This ticket allows clients to pass a list of hosts that will be used in
place of the default "first available host" behavior.

### Task Colocation

The prior system for finding nodes to execute a task worked worked only
with unassigned nodes. Any node assigned a task could not be assigned
another task.

This ticket adds a more capable prioritizer class that enables clients
using hostnames to colocate tasks. It also retains the capability to
return open nodes when no hostname is specified.