Unit tests for ShellLauncher & additional shell launch context (#671)
This PR merges in additional context to the ShellLauncher.start function to open a subprocess. The working directory is now set in popen, the environment variables are now set in popen, and the file paths to which standard output and standard error streams should be redirected to have been set. Additionally this PR merges in Unit tests for the ShellLauncher class.

[ reviewed by @MattToast ]
[ committed by @amandarichardsonn ]