Add patch to Tensorflow backend for GCC 14 compatibility (#738)
In libtensorflow, the `input` argument to `TF_SessionRun` seems to be
mistyped to `TF_Output` instead of `TF_Input`. These two types differ
only in name. GCC-14 catches this and throws an error, even though
earlier versions allow this. To solve this problem, patches are applied
to the Tensorflow backend in RedisAI. Future versions of Tensorflow may
fix this problem, but for now this seems to be the best workaround.

[ committed by @ashao ]
[ reviewed by @MattToast ]