Bypass executable check when creating RunSettings
In some instances, the strong check that the executable can be
found by the driver script when creating RunSettings is too
restrictive. This can occur in cases where the binary does not
exist on the node where the driver script is being launched (e.g.
remote execution) or for binaries which need to be found after
setting environment variables.